+++ stats/client/src/main/java/ru/practicum/client/BaseClient.java
@@ -1,4 +1,4 @@
-package ru.practicum;
+package ru.practicum.client;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.*;
