+++ stats/client/src/main/java/ru/practicum/client/StatsClient.java
@@ -1,4 +1,4 @@
-package ru.practicum;
+package ru.practicum.client;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -6,7 +6,7 @@ import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.util.DefaultUriBuilderFactory;
-import ru.practicum.dto.EndpointHitDto;
+import ru.practicum.models.dto.EndpointHitDto;
 
 import java.util.Map;
