+++ stats/models/src/main/java/ru/practicum/models/dto/EndpointHitDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.dto;
+package ru.practicum.models.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.PastOrPresent;
