+++ stats/models/src/main/java/ru/practicum/models/dto/ViewStats.java
@@ -1,4 +1,4 @@
-package ru.practicum.dto;
+package ru.practicum.models.dto;
 
 import lombok.*;
