+++ stats/service/src/main/java/ru/practicum/service/ExploreWithMeStatsService.java
@@ -1,4 +1,4 @@
-package ru.practicum;
+package ru.practicum.service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
